+++ code/go/0chain.net/smartcontract/dbs/event/event_test.go
@@ -38,42 +38,3 @@ func TestSetupDatabase(t *testing.T) {
 	err = eventDb.AutoMigrate()
 	require.NoError(t, err)
 }
-
-func (edb *EventDb) drop() error {
-	err := edb.Store.Get().Migrator().DropTable(&Event{})
-	if err != nil {
-		return err
-	}
-
-	err = edb.Store.Get().Migrator().DropTable(&Blobber{})
-	if err != nil {
-		return err
-	}
-
-	err = edb.Store.Get().Migrator().DropTable(&Transaction{})
-	if err != nil {
-		return err
-	}
-
-	err = edb.Store.Get().Migrator().DropTable(&WriteMarker{})
-	if err != nil {
-		return err
-	}
-
-	err = edb.Store.Get().Migrator().DropTable(&Validator{})
-	if err != nil {
-		return err
-	}
-
-	err = edb.Store.Get().Migrator().DropTable(&Block{})
-	if err != nil {
-		return err
-	}
-
-	err = edb.Store.Get().Migrator().DropTable(&ReadMarker{})
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
