+++ code/go/0chain.net/smartcontract/dbs/event/process.go
@@ -285,7 +285,7 @@ func (edb *EventDb) addStat(event Event) error {
 		if err != nil {
 			return err
 		}
-		return edb.addOrUpdateTotalMint(mint)
+		return edb.addOrUpdateTotalMint(&mint)
 	default:
 		return fmt.Errorf("unrecognised event %v", event)
 	}
