+++ code/go/0chain.net/smartcontract/dbs/event/transaction_test.go
@@ -41,26 +41,6 @@ func TestAddTransaction(t *testing.T) {
 	require.NoError(t, err)
 }
 
-func SetupDatabase(t *testing.T) *EventDb {
-
-	access := config.DbAccess{
-		Enabled:         true,
-		Name:            os.Getenv("POSTGRES_DB"),
-		User:            os.Getenv("POSTGRES_USER"),
-		Password:        os.Getenv("POSTGRES_PASSWORD"),
-		Host:            os.Getenv("POSTGRES_HOST"),
-		Port:            os.Getenv("POSTGRES_PORT"),
-		MaxIdleConns:    100,
-		MaxOpenConns:    200,
-		ConnMaxLifetime: 20 * time.Second,
-	}
-	eventDb, err := NewEventDb(access)
-	if err != nil {
-		t.Skip("only for local debugging, requires local postgresql")
-	}
-	return eventDb
-}
-
 func TestFindTransactionByHash(t *testing.T) {
 	access := config.DbAccess{
 		Enabled:         true,
